requirements.txt
transformers
faiss-cpu
numpy
torch
requests
httpx==0.23.0
openai
python-dotenv
llama-index
pinecone-client
google-cloud-translate
translate